[master] - done.

+++ RandomProduct/Models/Data/DiscountResult.cs
@@ -2,10 +2,10 @@
 {
     public class DiscountResult
     {
-        public int Percents { get; set; }
+        public double Percents { get; set; }
         public string ProductId { get; set; }
 
-        public DiscountResult(int percents, string productId)
+        public DiscountResult(double percents, string productId)
         {
             Percents = percents;
             ProductId = productId;
